test: split large index tests
